src/gpu: K6f Round 2 — small codegen-affecting hot-path nits (K6.33, K6.46, K6.47, K6.48)
K6.33 lz_decode_core.cuh:49, 80 — `match_len > MIN_PARALLEL_MATCH_LEN - 1`
  → `match_len >= MIN_PARALLEL_MATCH_LEN`. Mathematically identical;
  decode/lz_kernel.ptx byte-identical (verified via git diff).

K6.46 decode/huffman_kernel.cu — `int bit_count` / `int pending` →
  `uint32_t`. Both are always non-negative in this code (bit_count grows
  in REFILL_BITS chunks, drops by total_len ≤ MAX_CODE_LEN+1 which is
  always ≤ current value; pending grows by num_syms 1..2, drops by
  sizeof(uint32_t)=4 or 1). The `(int)sizeof(uint32_t)` casts at lines
  210/215 became `(uint32_t)sizeof(uint32_t)` (or removed) to match.

K6.47 encode/huffman_kernel.cu:68 — drop redundant
  `code & ((1u << len) - 1u)` mask. Canonical codes for length `len` are
  guaranteed to fit in `len` bits by construction (buildCanonicalCodes
  fills high bits with zero). Mask was a no-op.

K6.48 encode/huffman_kernel.cu:67 — add `assert(len > 0)` for the
  no-zero-length-code contract (buildCanonicalCodes preserves this for
  every used symbol). Catches malformed code-length tables at runtime in
  debug; nvcc strips with NDEBUG. Added `#include <cassert>`.

REG/STACK after Round 2 (vs Round 1):
  slzHuffEncode4StreamKernel REG:38 → REG:35   (small win from K6.47)
  slzHuffBuildLutKernel      unchanged
  slzHuffDecode4StreamKernel REG:40 unchanged  (PTX shifted but REG same)
  slzLzDecodeKernel          REG:40 unchanged  (PTX byte-identical)
  All other kernels: unchanged.

Verify:
  SHA enwik8 L1/L3/L5 roundtrip: all 2b49720e... ✓
  Bench L3 kernel best: 6.724 ms (post-K6e 6.714, baseline-band 6.687-6.694) — noise
  Bench L5 kernel best: 6.434 ms (post-K6e 6.420)                            — noise
  decode/lz_kernel.ptx: byte-identical (no LZ-decode hot-loop change)
  decode/huffman_kernel.ptx: shifted but REG unchanged, perf within noise
  encode/huffman_kernel.ptx: REG 38→35 (encoder-side win, not visible in
  decode bench)

+++ src/gpu/decode/huffman_kernel.cu
@@ -168,12 +168,12 @@ __device__ __forceinline__ uint32_t decodeStreamOneLane(
     uint8_t* __restrict__ out, uint32_t out_size)
 {
     uint64_t bit_buf = 0;
-    int bit_count = 0;
+    uint32_t bit_count = 0;
     uint32_t in_pos = 0;
     uint32_t out_pos = 0;     // total bytes decoded
     uint32_t written = 0;     // bytes flushed to global
     uint64_t acc = 0;         // pending decoded bytes (byte 0 = oldest)
-    int pending = 0;
+    uint32_t pending = 0;
 
     // Hot loop: 1 decode/iter, refill keeps >= 11 bits. Decoded bytes are
     // accumulated and flushed as 32-bit stores — the first <=3 flushes
@@ -207,12 +207,12 @@ __device__ __forceinline__ uint32_t decodeStreamOneLane(
         }
         bit_buf <<= total_len;
         bit_count -= total_len;
-        while (pending >= (int)sizeof(uint32_t)) {
+        while (pending >= sizeof(uint32_t)) {
             if (((uintptr_t)(out + written) & (alignof(uint32_t) - 1)) == 0u) {
                 *reinterpret_cast<uint32_t*>(out + written) = (uint32_t)acc;
                 acc >>= 32;
                 written  += (uint32_t)sizeof(uint32_t);
-                pending  -= (int)sizeof(uint32_t);
+                pending  -= (uint32_t)sizeof(uint32_t);
             } else {
                 out[written++] = (uint8_t)(acc & 0xFF);
                 acc >>= 8;
